+++ scripts/install.sh
@@ -6,6 +6,7 @@ if [ -x "$(which python)" ]; then
     virtualenv --clear --no-site-packages -p python py2
 
     ./py2/bin/pip install coverage
+    ./py2/bin/pip install https://github.com/conestack/node/archive/master.zip
     ./py2/bin/pip install -e .[test]
 fi
 if [ -x "$(which python3)" ]; then
@@ -14,5 +15,6 @@ if [ -x "$(which python3)" ]; then
     virtualenv --clear --no-site-packages -p python3 py3
 
     ./py3/bin/pip install coverage
+    ./py3/bin/pip install https://github.com/conestack/node/archive/master.zip
     ./py3/bin/pip install -e .[test]
 fi