+++ CHANGES.rst
@@ -5,10 +5,15 @@ History
 3.0.0 (unreleased)
 ------------------
 
+- Rename deprecated ``WidgetAttributes.allow_non_node_childs`` to
+  ``allow_non_node_children``.
+  [rnix]
+
 - Add ``push_state`` and ``pop_state`` to factory. Used in ``YafowilTestCase``
   for saving and restoring factory state. This is needed when yafowil tests run
   in compound with foreign tests to avoid breaking possible custom factory
   configuration.
+  [rnix]
 
 - Try to translate plain error message in ``error_renderer`` in case of no
   ``message_tag`` defined.